Move App files into dedicated folder
