+++ src/components/Search.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-import '../App.css';
 import Card from './Card';
 import Filters from './Filters';
