--- src/history.js
@@ -1,5 +1,0 @@
-import { createBrowserHistory } from 'history'
-
-export default createBrowserHistory({
-  /* pass a configuration object here if needed */
-})
